adjust creature API

+++ src/data/creature.py
@@ -1,5 +1,6 @@
-from .init import curs
+from .init import curs, IntegrityError
 from model.creature import Creature
+from errors import Missing, Duplicate
 
 curs.execute("""
     create table if not exists creature(
@@ -13,8 +14,8 @@ curs.execute("""
 
 
 def row_to_model(row: tuple) -> Creature:
-    name, description, country, area, aka = row
-    return Creature(name=name, description=description, country=country, area=area, aka=aka)
+    name, country, area, description, aka = row
+    return Creature(name=name, country=country, area=area, description=description, aka=aka)
 
 
 def model_to_dict(creature: Creature) -> dict | None:
@@ -26,7 +27,10 @@ def get_one(name: str) -> Creature:
     params = {'name': name}
     curs.execute(query, params)
     row = curs.fetchone()
-    return row_to_model(row)
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f'Creature {name} not found')
 
 
 def get_all() -> list[Creature]:
@@ -37,33 +41,49 @@ def get_all() -> list[Creature]:
 
 
 def create(creature: Creature) -> Creature:
+    if not creature:
+        return None
+
     query = """
-        insert into creature values
-        (:name, :description, :country, :area, :aka)
+        insert into creature (name, country, description, area, aka) 
+        values (:name, :country, :description, :area, :aka)
     """
     params = model_to_dict(creature)
-    curs.execute(query, params)
+    try:
+        curs.execute(query, params)
+    except IntegrityError:
+        raise Duplicate(msg=f'Creature {creature.name} already exists')
     return get_one(creature.name)
 
 
-def modify(creature: Creature) -> Creature:
+def modify(name: str, creature: Creature) -> Creature:
+    if not (name and creature):
+        return None
+
     query = """
         update creature
         set country=:country,
             name=:name,
-            description=:description,
-            area=:area,
-            aka=:aka
+            description=:description
+            area=:area
+            aka=:ala
         where name=:name_orig
     """
     params = model_to_dict(creature)
     params["name_orig"] = creature.name
-    _ = curs.execute(query, params)
-    return get_one(creature.name)
+    curs.execute(query, params)
+    if curs.rowcount == 1:
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f'Creature {name} not found')
 
 
-def delete(creature: Creature) -> bool:
+def delete(name: str) -> bool:
+    if not name:
+        return None
+
     query = 'delete from creature where name=:name'
-    params = {'name': creature.name}
-    res = curs.execute(query, params)
-    return bool(res)
+    params = {'name': name}
+    curs.execute(query, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f'Creature {name} not found')
